Rename letter page, path to Take Action
